Fix Random#rand(Range(Float, Float)) to return Float (#6445)
* Fix Random#rand(Range(Float, Float)) to return Float

* fixup! Fix Random#rand(Range(Float, Float)) to return Float

+++ src/random.cr
@@ -273,12 +273,12 @@ module Random
     rand_range(range)
   end
 
-  # Returns a random `Float64` in the given *range*.
+  # Returns a random `Float` in the given *range*.
   #
   # ```
   # Random.new.rand(6.2..21.768) # => 15.2989
   # ```
-  def rand(range : Range(Float, Float)) : Float64
+  def rand(range : Range(Float, Float)) : Float
     span = range.end - range.begin
     if range.excludes_end?
       unless range.begin < range.end
